QMAPS-2226 cleanup
