+++ src/panel/poi/PoiPanelContent.jsx
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import Telemetry from 'src/libs/telemetry';
 import ActionButtons from './ActionButtons';
 import PoiBlockContainer from './PoiBlockContainer';
-import OsmContribution from 'src/components/OsmContribution';
+import Contribution from 'src/components/Contribution';
 import CategoryList from 'src/components/CategoryList';
-import { isFromPagesJaunes, isFromOSM } from 'src/libs/pois';
+import { isFromPagesJaunes } from 'src/libs/pois';
 import { fire } from 'src/libs/customEvents';
 import PoiItem from 'src/components/PoiItem';
 import { Flex, Divider } from 'src/components/ui';
@@ -94,7 +94,7 @@ const PoiPanelContent = ({ poi }) => {
             <img src="./statics/images/pj.svg" alt="PagesJaunes" width="80" height="16" />
           </div>
         )}
-        {isFromOSM(poi) && <OsmContribution poi={poi} />}
+        <Contribution poi={poi} />
         <Divider paddingTop={0} className="poi_panel__fullWidth" />
         <h3 className="u-text--smallTitle u-mb-s">{_('Search around this place', 'poi')}</h3>
         <CategoryList className="poi_panel__categories u-mb-s" limit={4} />
