+++ src/scss/includes/panels/poi_panel.scss
@@ -1,4 +1,4 @@
-@import '../osm_contribute';
+@import '../contribute';
 
 $BLOCK_PADDING: 24px;
 $BLOCK_ICON_FONT_SIZE: 16px;
